pyproject.toml
[tool.poetry]
name = "movie-predict"
version = "0.1.0"
description = ""
authors = ["Volodymyr_Borysiuk <volodymyr.borysiuk0@gmail.com>"]
readme = "README.md"
packages = [{include = "movie_predict"}]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.23.5"
iter-model = "^2.1.0"
urlx = "^1.0.1"
uvicorn = "^0.20.0"
fastapi = "^0.88.0"
sqlmodel = "^0.0.8"
pydantic = "^1.10.2"
psycopg2-binary = "^2.9.5"
pyyaml = "^6.0"
asyncpg = "^0.27.0"
alembic = "^1.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
